Se elimino del sidebar el hook y se creo una carpeta para hacer un hook como componente
